moved Jcr-related functions from the kernel-api into kernel-modeshape
