+++ fcrepo-kernel-modeshape/src/main/java/org/fcrepo/kernel/modeshape/utils/UncheckedPredicate.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package org.fcrepo.kernel.api.utils;
+package org.fcrepo.kernel.modeshape.utils;
 
-import java.util.function.Function;
+import java.util.function.Predicate;
+
 import javax.jcr.RepositoryException;
 
 import org.fcrepo.kernel.api.exception.RepositoryRuntimeException;
@@ -25,36 +26,35 @@ import org.fcrepo.kernel.api.exception.RepositoryRuntimeException;
  * Operations that throw {@link RepositoryException} cannot be used as lambdas without "unchecking" those exceptions.
  *
  * @author ajs6f
- * @param <T> the type of the input to the function
+ * @param <T> the type of the input to the predicate
  */
 @FunctionalInterface
-public interface UncheckedFunction<T, R> extends Function<T, R> {
+public interface UncheckedPredicate<T> extends Predicate<T> {
 
     @Override
-    default R apply(final T elem) {
+    default boolean test(final T elem) {
         try {
-            return applyThrows(elem);
+            return testThrows(elem);
         } catch (final RepositoryException e) {
             throw new RepositoryRuntimeException(e);
         }
     }
 
     /**
-     * The same semantic as {@link #apply(Object)}, but allowed to throw a {@link RepositoryException}
+     * The same semantic as {@link #test(Object)}, but allowed to throw a {@link RepositoryException}
      *
      * @param elem the input argument
-     * @return the function result
-     * @throws RepositoryException the underlying repository error
+     * @return true if the input matches the predicate, otherwise false
+     * @throws RepositoryException a repository-related exception
      */
-    R applyThrows(T elem) throws RepositoryException;
+    boolean testThrows(T elem) throws RepositoryException;
 
     /**
-     * @param <T> the type of the input to the function
-     * @param <R> the type of the output of the function
+     * @param <T> the type of the input to the predicate
      * @param p a lambda expression
      * @return an unchecked version of that lambda
      */
-    static <T, R> UncheckedFunction<T, R> uncheck(final UncheckedFunction<T, R> p) {
+    static <T> UncheckedPredicate<T> uncheck(final UncheckedPredicate<T> p) {
         return p;
     }
 }
